Implement objection analysis in hippo routes
