+++ apps/legal_discovery/interface_flask.py
@@ -96,8 +96,7 @@ from .exhibit_routes import exhibits_bp
 from .trial_prep_routes import trial_prep_bp
 from .chain_logger import ChainEventType, log_event
 from .trial_assistant import bp as trial_assistant_bp  # noqa: E402
-from .hippo_routes import bp as hippo_bp  # noqa: E402
-from .objection_routes import bp as objections_bp  # noqa: E402
+from .hippo_routes import bp as hippo_bp, objections_bp  # noqa: E402
 from coded_tools.legal_discovery.bates_numbering import (
     BatesNumberingService,
     stamp_pdf,
