+++ apps/legal_discovery/trial_assistant/__init__.py
@@ -7,7 +7,6 @@ from ..database import db, log_retrieval_trace
 from ..models_trial import (
     TranscriptSegment,
     ObjectionEvent,
-    ObjectionResolution,
     TrialSession,
 )
 from .. import hippo
@@ -107,22 +106,4 @@ def handle_segment(data):
                 "trace_id": e.trace_id,
             },
             room=session_id,
-        )
-
-
-@socketio.on("objection_cure_chosen", namespace="/ws/trial")
-def cure_chosen(data):
-    evt_id = data.get("event_id")
-    cure = data.get("cure")
-    if not evt_id:
-        return
-    resolution = ObjectionResolution(event_id=evt_id, chosen_cure=cure)
-    db.session.add(resolution)
-    db.session.commit()
-    evt = db.session.get(ObjectionEvent, evt_id)
-    if evt:
-        emit(
-            "clear_highlights",
-            {"segment_id": evt.segment_id},
-            room="trial_objections",
         )
